requirements.txt
boto3
fastapi
python-dotenv
virtualenv